LayoutTests/platform/ios-sim-deprecated/editing/pasteboard/two-step-paste.html

<html>
<head>
<script src="../../../../resources/js-test-pre.js"></script>
<script src="../../../../editing/editing.js"></script>
</head>
<body>
    <div id="container" contenteditable="true">
        <p id="paragraph" style="color: rgb(153, 153, 153);">This is some text</p>
    </div>
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();

        description("The following test does a copy and a paste of contenteditable content. It then makes sure that the paste involved a two-step paste. The mechanism it uses to verify this is to determine if some styles change appropriately.");

        var container = document.getElementById("container");
        var selection = window.getSelection();
        var range = document.createRange();
        container.focus();
        range.selectNodeContents(container);
        selection.removeAllRanges();
        selection.addRange(range);

        copyCommand();
        pasteCommand();

        var t = document.getElementById("paragraph");
        shouldNotBe("t.style.color", "\"rgb(153, 153, 153)\"");
    </script>
<script src="../../../../resources/js-test-post.js"></script>
</html>
