safari-600.7-branch/LayoutTests/platform/ios-sim/editing/pasteboard/two-step-paste-expected.txt

The following test does a copy and a paste of contenteditable content. It then makes sure that the paste involved a two-step paste. The mechanism it uses to verify this is to determine if some styles change appropriately.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS t.style.color is not "rgb(153, 153, 153)"
PASS successfullyParsed is true

TEST COMPLETE
This is some text
